Mercredi
en cours: type enemy, type tower

+++ Assets/Scripts/ScriptableObject/EnemyData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyData", menuName = "Scriptable Objects/EnemyData")]
 public class EnemyData : ScriptableObject
 {
-    public int life;
-    public int damage;
-    public int speed;
+    public float life;
+    public float damage;
+    public float speed;
 }
